refactor(ccyl): extract model classes from ccyl_service.dart to models/ccyl_models.dart
