+++ lib/providers/ccyl_provider.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/foundation.dart';
+import 'package:bugaoshan/pages/campus/ccyl/models/ccyl_models.dart';
 import 'package:bugaoshan/services/api/ccyl_api_service.dart';
 import 'package:bugaoshan/services/auth/ccyl_auth.dart';
-import 'package:bugaoshan/services/ccyl/ccyl_service.dart';
 
 /// 第二课堂（CCYL）登录状态的 Provider。
 ///
